requirements.txt
Pillow==9.4.0
Flask==2.2.3
Flask-Cors==3.0.10
torch==2.0.0
transformers==4.27.4